Task 1/requirements.txt
google-generativeai>=0.3.0,<1.0.0
langchain>=0.1.0,<1.0.0
langchain-google-genai>=0.0.1,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
requests>=2.25.0,<3.0.0
jinja2>=3.0.0,<4.0.0 